Add Bower package support
As it is, you can clone the repository, or download external.js and add
it to your project. However, it'd be great to be able to use this
package via Bower, so you can add it as a dependency in a project's
bower.json file, or search it via Bower.io. Given that, I'm adding
Bower package support to the repository.

For further details about Bower package support, check out the official
documentation at https://bower.io/docs/creating-packages/.